// src/generated/api/models/ResponseDtoString.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ResponseDtoString
 */
export interface ResponseDtoString {
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoString
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoString
     */
    message?: string;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoString
     */
    data?: string;
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoString
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDtoString interface.
 */
export function instanceOfResponseDtoString(value: object): value is ResponseDtoString {
    return true;
}

export function ResponseDtoStringFromJSON(json: any): ResponseDtoString {
    return ResponseDtoStringFromJSONTyped(json, false);
}

export function ResponseDtoStringFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDtoString {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : json['data'],
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ResponseDtoStringToJSON(json: any): ResponseDtoString {
    return ResponseDtoStringToJSONTyped(json, false);
}

export function ResponseDtoStringToJSONTyped(value?: ResponseDtoString | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'message': value['message'],
        'data': value['data'],
        'timestamp': value['timestamp'],
    };
}