// src/generated/api/models/ResponseDtoUserActivityResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserActivityResponse } from './UserActivityResponse';
import {
    UserActivityResponseFromJSON,
    UserActivityResponseFromJSONTyped,
    UserActivityResponseToJSON,
    UserActivityResponseToJSONTyped,
} from './UserActivityResponse';

/**
 * 
 * @export
 * @interface ResponseDtoUserActivityResponse
 */
export interface ResponseDtoUserActivityResponse {
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoUserActivityResponse
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoUserActivityResponse
     */
    message?: string;
    /**
     * 
     * @type {UserActivityResponse}
     * @memberof ResponseDtoUserActivityResponse
     */
    data?: UserActivityResponse;
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoUserActivityResponse
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDtoUserActivityResponse interface.
 */
export function instanceOfResponseDtoUserActivityResponse(value: object): value is ResponseDtoUserActivityResponse {
    return true;
}

export function ResponseDtoUserActivityResponseFromJSON(json: any): ResponseDtoUserActivityResponse {
    return ResponseDtoUserActivityResponseFromJSONTyped(json, false);
}

export function ResponseDtoUserActivityResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDtoUserActivityResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : UserActivityResponseFromJSON(json['data']),
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ResponseDtoUserActivityResponseToJSON(json: any): ResponseDtoUserActivityResponse {
    return ResponseDtoUserActivityResponseToJSONTyped(json, false);
}

export function ResponseDtoUserActivityResponseToJSONTyped(value?: ResponseDtoUserActivityResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'message': value['message'],
        'data': UserActivityResponseToJSON(value['data']),
        'timestamp': value['timestamp'],
    };
}