// src/generated/api/models/ResponseDtoUserProfileResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserProfileResponse } from './UserProfileResponse';
import {
    UserProfileResponseFromJSON,
    UserProfileResponseFromJSONTyped,
    UserProfileResponseToJSON,
    UserProfileResponseToJSONTyped,
} from './UserProfileResponse';

/**
 * 
 * @export
 * @interface ResponseDtoUserProfileResponse
 */
export interface ResponseDtoUserProfileResponse {
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoUserProfileResponse
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ResponseDtoUserProfileResponse
     */
    message?: string;
    /**
     * 
     * @type {UserProfileResponse}
     * @memberof ResponseDtoUserProfileResponse
     */
    data?: UserProfileResponse;
    /**
     * 
     * @type {number}
     * @memberof ResponseDtoUserProfileResponse
     */
    timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDtoUserProfileResponse interface.
 */
export function instanceOfResponseDtoUserProfileResponse(value: object): value is ResponseDtoUserProfileResponse {
    return true;
}

export function ResponseDtoUserProfileResponseFromJSON(json: any): ResponseDtoUserProfileResponse {
    return ResponseDtoUserProfileResponseFromJSONTyped(json, false);
}

export function ResponseDtoUserProfileResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDtoUserProfileResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : json['status'],
        'message': json['message'] == null ? undefined : json['message'],
        'data': json['data'] == null ? undefined : UserProfileResponseFromJSON(json['data']),
        'timestamp': json['timestamp'] == null ? undefined : json['timestamp'],
    };
}

export function ResponseDtoUserProfileResponseToJSON(json: any): ResponseDtoUserProfileResponse {
    return ResponseDtoUserProfileResponseToJSONTyped(json, false);
}

export function ResponseDtoUserProfileResponseToJSONTyped(value?: ResponseDtoUserProfileResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': value['status'],
        'message': value['message'],
        'data': UserProfileResponseToJSON(value['data']),
        'timestamp': value['timestamp'],
    };
}